+++ Source/Core/DiscIO/VolumeWad.h
@@ -30,7 +30,7 @@ public:
 	bool GetTitleID(u8* _pBuffer) const override;
 	std::string GetUniqueID() const override;
 	std::string GetMakerID() const override;
-	int GetRevision() const override;
+	u16 GetRevision() const override;
 	std::string GetInternalName() const override { return ""; }
 	std::map<IVolume::ELanguage, std::string> GetNames() const override;
 	u32 GetFSTSize() const override { return 0; }
